Adding new src files
apiWrapper is a simple wrapper around the curlpp library and Scryfall's
API. It's used mostly to parse the calls to the website into files that
can be used later by other classes.

card is the backbone of the project. It's the smallest unit of the Magic
collection. It's very simple at the moment but changes will be added in
the future to enable more features and better information control.

collection is a class that's designed to keep track of all the cards and
decks that the user has. It will offer a simple but effective interface
to show information about all the cards.